requirements.txt
django
python-telegram-bot==13.11
lxml
beautifulsoup4
requests
djangorestframework
django-after-response